requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.23
tqdm>=4.66
Pillow>=10.0
matplotlib>=3.7
pyyaml>=6.0
pandas>=2.0

transformers>=4.40
accelerate>=0.30
safetensors>=0.4

pycocotools>=2.0.7
tensorboard>=2.14

diffusers>=0.30
huggingface_hub>=0.24
opencv-python>=4.8
scipy>=1.10

einops>=0.7
